base upload class
