+++ src/jigsaw/component/upload/upload.html
@@ -7,7 +7,7 @@
         <span class="jigsaw-upload-prompt">将文件拖拽到此处上传</span>
         <j-button (click)="_$selectFile($event)">选择文件</j-button>
         <input #fileInput class="jigsaw-upload-file-input" type='file' [accept]="fileType"
-               [multiple]="multiple" (change)="_$selectedFileChange()">
+               [multiple]="multiple">
     </j-box>
 
     <j-box *ngSwitchCase="'single'" type="flex" direction="v" justify="center" align="center" width="100%"
